requirements.txt
#flask
#requests
#flask-bootstrap
#flask-restplus
#flask-swagger-ui
#azure-functions
#azure-cosmos
#pymongo
#azure-cli-core
#feedgen

requests==2.23.0
Flask-Bootstrap==3.3.7.1
flask-restplus==0.13.0
flask-swagger-ui==3.36.0
#werkzeug
#azure-functions==1.7.2
azure-cosmos
pymongo
#azure-cli-core = "*"
feedgen==0.9.0
Flask==1.1